♻️ refactor: refactor and clean global store and chat store (#774)
* ♻️ refactor: refactor the globalStore implement and usage

* 🔥 refactor: clean deprecated code

* ♻️ refactor: refactor the chat store to slice mode
