+++ src/features/AgentSetting/AgentMeta/index.tsx
@@ -8,7 +8,8 @@ import { memo, useMemo } from 'react';
 import { useTranslation } from 'react-i18next';
 
 import { FORM_STYLE } from '@/const/layoutTokens';
-import { settingsSelectors, useGlobalStore } from '@/store/global';
+import { useGlobalStore } from '@/store/global';
+import { settingsSelectors } from '@/store/global/selectors';
 
 import { useStore } from '../store';
 import { SessionLoadingState } from '../store/initialState';
