+++ src/store/chat/slices/enchance/action.test.ts
@@ -1,4 +1,4 @@
-import { act, renderHook, waitFor } from '@testing-library/react';
+import { act, renderHook } from '@testing-library/react';
 import { Mock, afterEach, beforeEach, describe, expect, it, vi } from 'vitest';
 
 import { chainLangDetect } from '@/chains/langDetect';
@@ -6,7 +6,7 @@ import { chainTranslate } from '@/chains/translate';
 import { chatService } from '@/services/chat';
 import { messageService } from '@/services/message';
 
-import { useChatStore } from '../store';
+import { useChatStore } from '../../store';
 
 // Mock messageService 和 chatService
 vi.mock('@/services/message', () => ({
