+++ src/store/chat/slices/share/initialState.ts
@@ -1,0 +1,7 @@
+export interface ChatShareState {
+  shareLoading?: boolean;
+}
+
+export const initialShareState: ChatShareState = {
+  shareLoading: false,
+};
