+++ src/store/chat/slices/tool/initialState.ts
@@ -1,0 +1,7 @@
+export interface ChatToolState {
+  dalleImageLoading: Record<string, boolean>;
+}
+
+export const initialToolState: ChatToolState = {
+  dalleImageLoading: {},
+};
