+++ src/store/global/selectors.ts
@@ -1,2 +1,2 @@
-export * from './selectors/preference';
-export * from './selectors/settings';
+export * from './slices/common/selectors';
+export * from './slices/settings/selectors';
