+++ src/store/global/slices/common/selectors.ts
@@ -1,6 +1,6 @@
 import { GlobalStore } from '@/store/global';
 
-import { initialState } from '../initialState';
+import { initialState } from '../../initialState';
 
 const sessionGroupKeys = (s: GlobalStore): string[] =>
   s.preference.sessionGroupKeys || initialState.preference.sessionGroupKeys;
