+++ src/store/global/slices/settings/selectors.test.ts
@@ -1,7 +1,7 @@
 import { LanguageModel } from '@/types/llm';
 
-import { GlobalStore } from '../store';
-import { settingsSelectors } from './settings';
+import { GlobalStore } from '../../store';
+import { settingsSelectors } from './selectors';
 
 describe('settingsSelectors', () => {
   describe('currentSettings', () => {
